@@ -1,15 +1,13 @@
-import uuid
 from datetime import datetime
 from enum import Enum
-from typing import Optional, List
 
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 
-from app.core.types import BaseIdType
-from app.schemas.card import CardStatus
+from app.core.custom_types import BaseIdType
 
 
-class SessionSettings(BaseModel): ...
+class BaseSession(BaseModel):
+    pass
 
 
 class SessionMode(str, Enum):
@@ -23,23 +21,31 @@ class SessionStatus(str, Enum):
     ABANDONED = "abandoned"
 
 
-class SessionCreate(BaseModel):
-    roadmap_id: BaseIdType
-    block_id: Optional[BaseIdType] = None
+class SessionCreate(BaseSession):
     mode: SessionMode
-    settings: SessionSettings = SessionSettings()
+    roadmap_id: BaseIdType
+    block_id: BaseIdType | None = None
+    mix: bool | None = False
 
 
-class SessionInDB(BaseModel):
+class SessionUpdate(BaseModel):
+    status: SessionStatus | None = None
+    current_card_index: int | None = None
+    correct_answers: int | None = None
+    incorrect_answers: int | None = None
+    review_answers: int | None = None
+
+
+class SessionRead(BaseSession):
     id: BaseIdType
     user_id: BaseIdType
     roadmap_id: BaseIdType
-    block_id: Optional[BaseIdType] = None
+    block_id: BaseIdType | None = None
 
     mode: SessionMode
 
     status: SessionStatus
-    card_queue: List[BaseIdType] = []
+    card_ids_queue: list[BaseIdType] = []
     current_card_index: int
 
     correct_answers: int
@@ -48,34 +54,24 @@ class SessionInDB(BaseModel):
 
     created_at: datetime
     updated_at: datetime
-    completed_at: Optional[datetime] = None
+    completed_at: datetime | None = None
 
-    class Config:
-        from_attributes = True
-
-
-class SessionResponse(SessionInDB): ...
-
-
-class SessionUpdate(BaseModel):
-    status: Optional[SessionStatus] = None
-    current_card_index: Optional[int] = None
-    correct_answers: Optional[int] = None
-    incorrect_answers: Optional[int] = None
-    review_answers: Optional[int] = None
+    model_config = ConfigDict(from_attributes=True)
 
 
 class SessionFilters(BaseModel):
-    roadmap_id: Optional[BaseIdType] = None
-    block_id: Optional[uuid.UUID] = None
-    mode: Optional[SessionMode] = None
-    status: Optional[SessionStatus] = None
+    user_id: BaseIdType | None = None
+    mode: SessionMode | None = None
+    roadmap_id: BaseIdType | None = None
+    block_id: BaseIdType | None = None
+    status: SessionStatus | None = None
 
 
 class SessionResult(BaseModel):
     id: BaseIdType
     user_id: BaseIdType
     roadmap_id: BaseIdType
+    block_id: BaseIdType | None = None
     mode: SessionMode
     total_cards: int
     correct_answers: int
@@ -83,8 +79,3 @@ class SessionResult(BaseModel):
     review_answers: int
     accuracy_percentage: float
     completed_at: datetime
-
-
-class SubmitAnswerRequest(BaseModel):
-    card_id: BaseIdType
-    answer: CardStatus
