@@ -38,7 +38,7 @@ def upgrade() -> None:
         type_=sa.String(length=1024),
         existing_nullable=False,
     )
-    op.drop_constraint(op.f("users_email_key"), "users", type_="unique")
+    # op.drop_constraint(op.f("users_email_key"), "users", type_="unique")
     op.create_index(op.f("ix_users_email"), "users", ["email"], unique=True)
     # ### end Alembic commands ###
 
@@ -47,9 +47,9 @@ def downgrade() -> None:
     """Downgrade schema."""
     # ### commands auto generated by Alembic - please adjust! ###
     op.drop_index(op.f("ix_users_email"), table_name="users")
-    op.create_unique_constraint(
-        op.f("users_email_key"), "users", ["email"], postgresql_nulls_not_distinct=False
-    )
+    # op.create_unique_constraint(
+    #     op.f("users_email_key"), "users", ["email"], postgresql_nulls_not_distinct=False
+    # )
     op.alter_column(
         "users",
         "hashed_password",
