@@ -42,21 +42,21 @@ def upgrade() -> None:
         ["created_at"],
         unique=False,
     )
-    op.drop_constraint(op.f("users_username_key"), "users", type_="unique")
-    op.create_unique_constraint(op.f("uq_users_username"), "users", ["username"])
+    # op.drop_constraint(op.f("users_username_key"), "users", type_="unique")
+    # op.create_unique_constraint(op.f("uq_users_username"), "users", ["username"])
     # ### end Alembic commands ###
 
 
 def downgrade() -> None:
     """Downgrade schema."""
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f("uq_users_username"), "users", type_="unique")
-    op.create_unique_constraint(
-        op.f("users_username_key"),
-        "users",
-        ["username"],
-        postgresql_nulls_not_distinct=False,
-    )
+    # op.drop_constraint(op.f("uq_users_username"), "users", type_="unique")
+    # op.create_unique_constraint(
+    #     op.f("users_username_key"),
+    #     "users",
+    #     ["username"],
+    #     postgresql_nulls_not_distinct=False,
+    # )
     op.drop_index(op.f("ix_access_tokens_created_at"), table_name="access_tokens")
     op.drop_table("access_tokens")
     # ### end Alembic commands ###
