@@ -6,11 +6,12 @@ from fastapi.security import HTTPBearer
 
 from app.core.config import settings
 from .auth import router as auth_router
+from .user import router as user_router
+from .roadmap import router as roadmap_router
 from .block import router as block_router
 from .card import router as card_router
-from .roadmap import router as roadmap_router
-from .session import router as session_manager_router
-from .user import router as user_router
+
+from .session import router as session_router
 
 http_bearer = HTTPBearer(auto_error=False)
 
@@ -24,4 +25,4 @@ router.include_router(user_router)
 router.include_router(roadmap_router)
 router.include_router(block_router)
 router.include_router(card_router)
-router.include_router(session_manager_router)
+router.include_router(session_router)
