@@ -7,7 +7,7 @@ from app.core.authentication.fastapi_users import current_active_user
 from app.core.config import settings
 from app.core.dependencies.services import get_block_service
 from app.core.handlers import router_handler
-from app.core.types import BaseIdType
+from app.core.custom_types import BaseIdType
 from app.schemas.block import (
     BlockRead,
     BlockCreate,
@@ -93,7 +93,7 @@ async def get_block(
 
 # -------------------------------------- CREATE --------------------------------------
 @router.post(
-    "/",
+    "",
     name="blocks:create_block",
     response_model=BlockRead,
 )
