@@ -2,7 +2,7 @@ from fastapi_users import FastAPIUsers
 
 from app.core.dependencies import authentication_backend
 from app.core.dependencies.users import get_user_manager
-from app.core.types import BaseIdType
+from app.core.custom_types import BaseIdType
 from app.models import User
 
 fastapi_users = FastAPIUsers[User, BaseIdType](
