@@ -2,38 +2,32 @@ from typing import Annotated, TYPE_CHECKING
 
 from fastapi import Depends
 
-from app.services import AccessService, RoadmapService
-from app.services import BlockService, CardService, SessionManagerService, UserService
+from app.services import (
+    UserService,
+    RoadmapService,
+    BlockService,
+    CardService,
+    SessionService,
+)
 from .repositories import (
     get_user_repository,
     get_roadmap_repository,
     get_block_repository,
     get_card_repository,
-    get_session_manager_repository,
+    get_session_repository,
 )
 
+from .cache import get_redis
+
 if TYPE_CHECKING:
+    from redis.asyncio import Redis
     from repositories import (
         UserRepository,
         RoadmapRepository,
-        CardRepository,
         BlockRepository,
-        SessionManagerRepository,
+        CardRepository,
+        SessionRepository,
     )
-    from repositories import BlockRepository
-
-
-async def get_access_service(
-    roadmap_repo: Annotated[
-        "RoadmapRepository",
-        Depends(get_roadmap_repository),
-    ],
-    block_repo: Annotated[
-        "BlockRepository",
-        Depends(get_block_repository),
-    ],
-) -> AccessService:
-    yield AccessService(roadmap_repo, block_repo)
 
 
 async def get_user_service(
@@ -41,12 +35,15 @@ async def get_user_service(
         "UserRepository",
         Depends(get_user_repository),
     ],
-    access_service: Annotated[
-        "AccessService",
-        Depends(get_access_service),
+    redis: Annotated[
+        "Redis",
+        Depends(get_redis),
     ],
 ) -> UserService:
-    yield UserService(user_repo, access_service)
+    yield UserService(
+        user_repo,
+        redis,
+    )
 
 
 async def get_roadmap_service(
@@ -54,12 +51,15 @@ async def get_roadmap_service(
         "RoadmapRepository",
         Depends(get_roadmap_repository),
     ],
-    access_service: Annotated[
-        "AccessService",
-        Depends(get_access_service),
+    redis: Annotated[
+        "Redis",
+        Depends(get_redis),
     ],
 ) -> RoadmapService:
-    yield RoadmapService(repo, access_service)
+    yield RoadmapService(
+        repo,
+        redis,
+    )
 
 
 async def get_block_service(
@@ -67,12 +67,15 @@ async def get_block_service(
         "BlockRepository",
         Depends(get_block_repository),
     ],
-    access_service: Annotated[
-        "AccessService",
-        Depends(get_access_service),
+    redis: Annotated[
+        "Redis",
+        Depends(get_redis),
     ],
 ) -> BlockService:
-    yield BlockService(repo, access_service)
+    yield BlockService(
+        repo,
+        redis,
+    )
 
 
 async def get_card_service(
@@ -80,18 +83,28 @@ async def get_card_service(
         "CardRepository",
         Depends(get_card_repository),
     ],
-    access_service: Annotated[
-        "AccessService",
-        Depends(get_access_service),
+    redis: Annotated[
+        "Redis",
+        Depends(get_redis),
     ],
 ) -> CardService:
-    yield CardService(repo, access_service)
+    yield CardService(
+        repo,
+        redis,
+    )
 
 
-async def get_session_manager_service(
+async def get_session_service(
     repo: Annotated[
-        "SessionManagerRepository",
-        Depends(get_session_manager_repository),
+        "SessionRepository",
+        Depends(get_session_repository),
     ],
-) -> SessionManagerService:
-    yield SessionManagerService(repo)
+    redis: Annotated[
+        "Redis",
+        Depends(get_redis),
+    ],
+) -> SessionService:
+    yield SessionService(
+        repo,
+        redis,
+    )
