@@ -5,7 +5,7 @@ from fastapi_users import exceptions
 from sqlalchemy import ForeignKey, DateTime, func, text
 from sqlalchemy.orm import declared_attr, Mapped, mapped_column, relationship
 
-from app.core.types import BaseIdType
+from app.core.custom_types import BaseIdType
 from app.utils.generators import server_id_generator, id_generator
 
 if TYPE_CHECKING:
@@ -17,7 +17,7 @@ if TYPE_CHECKING:
 class IdMixin:
     id: Mapped[BaseIdType] = mapped_column(
         primary_key=True,
-        default=id_generator(),
+        default=id_generator,
         server_default=text(server_id_generator()),
     )
 
@@ -28,27 +28,6 @@ class IdMixin:
             return BaseIdType(value)
         except ValueError as e:
             raise exceptions.InvalidID() from e
-
-
-class UserRelationMixin:
-    _user_id_nullable: bool = False
-    _user_id_unique: bool = False
-    _user_back_populates: str | None = None
-
-    @declared_attr
-    def user_id(cls) -> Mapped[BaseIdType]:
-        return mapped_column(
-            ForeignKey("users.id"),
-            unique=cls._user_id_unique,
-            nullable=cls._user_id_nullable,
-        )
-
-    @declared_attr
-    def user(cls) -> Mapped["User"]:
-        return relationship(
-            "User",
-            back_populates=cls._user_back_populates,
-        )
 
 
 class TimestampMixin:
@@ -63,6 +42,27 @@ class TimestampMixin:
     )
 
 
+class UserRelationMixin:
+    _user_id_nullable: bool = False
+    _user_id_unique: bool = False
+    _user_back_populates: str | None = None
+
+    @declared_attr
+    def user_id(cls) -> Mapped[BaseIdType]:
+        return mapped_column(
+            ForeignKey("users.id", ondelete="CASCADE"),
+            unique=cls._user_id_unique,
+            nullable=cls._user_id_nullable,
+        )
+
+    @declared_attr
+    def user(cls) -> Mapped["User"]:
+        return relationship(
+            "User",
+            back_populates=cls._user_back_populates,
+        )
+
+
 class RoadmapRelationMixin:
     _roadmap_id_nullable: bool = False
     _roadmap_id_unique: bool = False
@@ -71,7 +71,7 @@ class RoadmapRelationMixin:
     @declared_attr
     def roadmap_id(cls) -> Mapped[BaseIdType]:
         return mapped_column(
-            ForeignKey("roadmaps.id"),
+            ForeignKey("roadmaps.id", ondelete="CASCADE"),
             unique=cls._roadmap_id_unique,
             nullable=cls._roadmap_id_nullable,
         )
@@ -92,7 +92,7 @@ class BlockRelationMixin:
     @declared_attr
     def block_id(cls) -> Mapped[BaseIdType]:
         return mapped_column(
-            ForeignKey("blocks.id"),
+            ForeignKey("blocks.id", ondelete="CASCADE"),
             unique=cls._block_id_unique,
             nullable=cls._block_id_nullable,
         )
