@@ -1,10 +1,11 @@
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
 
+from app.models import Base
+
 if TYPE_CHECKING:
     from sqlalchemy.ext.asyncio import AsyncSession
-    from app.core.types import BaseIdType
-    from app.models import Base
+    from app.core.custom_types import BaseIdType
 
 
 class BaseRepository(ABC):
@@ -12,19 +13,19 @@ class BaseRepository(ABC):
         self.session = session
 
     @abstractmethod
-    async def get_all(self) -> list["Base"]:
+    async def get_all(self) -> list[Base]:
         pass
 
     @abstractmethod
-    async def get_by_filters(self, filters: dict) -> list["Base"]:
+    async def get_by_filters(self, filters: dict) -> list[Base]:
         pass
 
     @abstractmethod
-    async def get_by_id(self, object_id: "BaseIdType") -> "Base" | None:
+    async def get_by_id(self, object_id: "BaseIdType") -> Base | None:
         pass
 
     @abstractmethod
-    async def create(self, create_data: dict) -> "Base" | None:
+    async def create(self, create_data: dict) -> Base | None:
         pass
 
     @abstractmethod
@@ -32,5 +33,5 @@ class BaseRepository(ABC):
         pass
 
     @abstractmethod
-    async def update(self, object_id: "BaseIdType", update_data: dict) -> "Base" | None:
+    async def update(self, object_id: "BaseIdType", update_data: dict) -> Base | None:
         pass
