@@ -3,7 +3,7 @@ from enum import Enum
 
 from pydantic import BaseModel, ConfigDict
 
-from app.core.types import BaseIdType
+from app.core.custom_types import BaseIdType
 
 
 class BaseBlock(BaseModel):
@@ -12,7 +12,7 @@ class BaseBlock(BaseModel):
 
 
 class BlockCreate(BaseBlock):
-    order_index: int
+    order_index: float
     roadmap_id: BaseIdType
 
 
@@ -26,14 +26,15 @@ class BlockUpdate(BaseModel):
     title: str | None = None
     description: str | None = None
     status: BlockStatus | None = None
-    order_index: int | None = None
+    order_index: float | None = None
     roadmap_id: BaseIdType | None = None
 
 
 class BlockRead(BaseBlock):
     id: BaseIdType
     roadmap_id: BaseIdType
-    order_index: int
+    user_id: BaseIdType
+    order_index: float
     status: BlockStatus
     created_at: datetime
     updated_at: datetime
@@ -42,8 +43,9 @@ class BlockRead(BaseBlock):
 
 
 class BlockFilters(BaseModel):
+    user_id: BaseIdType | None = None
     roadmap_id: BaseIdType | None = None
     title: str | None = None
     description: str | None = None
     status: BlockStatus | None = None
-    order_index: int | None = None
+    order_index: float | None = None
