@@ -1,19 +1,17 @@
 __all__ = (
     "BaseService",
-    "UserService",
     "UserManager",
+    "UserService",
     "RoadmapService",
     "BlockService",
     "CardService",
-    "SessionManagerService",
-    "AccessService",
+    "SessionService",
 )
 
-from .access import AccessService
 from .base import BaseService
+from .user_manager import UserManager
+from .user import UserService
+from .roadmap import RoadmapService
 from .block import BlockService
 from .card import CardService
-from .roadmap import RoadmapService
-from .session_manager import SessionManagerService
-from .user import UserService
-from .user_manager import UserManager
+from .session import SessionService
