@@ -1,11 +1,16 @@
+import json
+import logging
 from typing import TYPE_CHECKING
 
+from redis.asyncio import Redis
+
 from app.core.handlers import service_handler
-from app.core.logging import user_service_logger as logger
+from app.core.loggers import user_service_logger as logger
+from app.shared.access import get_accessed_filters
+from app.utils.mappers.cache_to_model import users_cache_to_model
 from app.utils.mappers.orm_to_models import user_orm_to_model
 
 if TYPE_CHECKING:
-    from app.services import AccessService
     from app.repositories import UserRepository
     from app.schemas.user import UserRead, UserFilters
     from app.models import User
@@ -15,20 +20,31 @@ class UserService:
     def __init__(
         self,
         repo: "UserRepository",
-        access_service: "AccessService",
+        redis: "Redis",
     ):
         self.repo = repo
-        self.access = access_service
+        self.redis = redis
+        self.ttl = 60
 
     @service_handler
     async def get_all(self) -> list["UserRead"]:
+        cache_key = "users:all"
+        cached = await self.redis.get(cache_key)
+        if cached:
+            logging.info("Hit cache for key: %r", cache_key)
+            return await users_cache_to_model(cached)
+
         db_users = await self.repo.get_all()
         if len(db_users) == 0:
             logger.warning("Users not found in DB")
             return []
 
         validated_users = [await user_orm_to_model(user) for user in db_users]
-
+        await self.redis.set(
+            cache_key,
+            json.dumps([u.model_dump_json() for u in validated_users]),
+            ex=self.ttl,
+        )
         return validated_users
 
     @service_handler
@@ -38,7 +54,7 @@ class UserService:
         filters: "UserFilters",
     ) -> list["UserRead"]:
         filters_dict = filters.model_dump()
-        accessed_filters = await self.access.filter_users_for_user(
+        accessed_filters = await get_accessed_filters(
             current_user,
             filters_dict,
         )
