@@ -1,20 +1,48 @@
+from datetime import datetime, timezone
 from typing import Optional, TYPE_CHECKING
 
 from fastapi_users import BaseUserManager
+from sqlalchemy import delete, and_, select, asc
 
 from app.core.config import settings
-from app.core.logging import user_manager_logger as logger
-from app.core.types import BaseIdType
-from app.models import User
+from app.core.loggers import user_manager_logger as logger
+from app.core.custom_types import BaseIdType
+from app.models import User, db_helper, AccessToken
 from app.models.mixins import IdMixin
 
 if TYPE_CHECKING:
-    from fastapi import Request
+    from fastapi import Request, Response
+    from app.models.user import SQLAlchemyUserDatabase
+    from app.models.access_token import SQLAlchemyAccessTokenDatabase
 
 
 class UserManager(IdMixin, BaseUserManager[User, BaseIdType]):
     reset_password_token_secret = settings.access_token.reset_password_token_secret
     verification_token_secret = settings.access_token.verification_token_secret
+
+    def __init__(
+        self,
+        user_db: "SQLAlchemyUserDatabase",
+        access_tokens_db: "SQLAlchemyAccessTokenDatabase",
+    ):
+        super().__init__(user_db)
+        self.access_tokens_db = access_tokens_db
+
+    async def on_after_login(
+        self,
+        user: User,
+        request: Optional["Request"] = None,
+        response: Optional["Response"] = None,
+    ) -> None:
+        logger.info("[on_after_login] user %r logged in", user.id)
+
+        result = await self.access_tokens_db.cleanup_user_tokens(user.id)
+
+        logger.info(
+            "[on_after_login] token cleanup result for user %r: %r",
+            user.id,
+            result,
+        )
 
     async def on_after_register(
         self,
