@@ -1,7 +1,7 @@
 import uuid
 
-from app.core.types import BaseIdType
+from app.core.custom_types import BaseIdType
 
 
-async def generate_base_id() -> BaseIdType:
+def generate_base_id() -> BaseIdType:
     return uuid.uuid4()
